Removed using of count_int_str_bits
Else utf8 mode is not properly wrapping since integers are handled differently than str

For example:

When passing "secret", it is decoded to 0x736563726574 (int).
Therefore, the bit_length of the number is used which is 47.

When passing "0x736563726574" (str), the function count_int_str_bits returns the amount of hex digits times 4 which is 48.

The encryption then runs on word with different size which leads to inconsistent ciphertext even though those inputs should lead to the same ciphertext.

This could be fixed by also using count_int_str_bits on the integer; casting it to a string beforehand.

I chose to remove count_int_str_bits because I don't know anymore why I introduced it in the first place...